Rename files, move row type definition into constants file, lint
